Move code block commands into code-block.ts
