+++ src/rich-text/commands/index.ts
@@ -4,13 +4,7 @@ import {
     toggleMark,
     wrapIn,
 } from "prosemirror-commands";
-import {
-    Mark,
-    MarkType,
-    NodeType,
-    Schema,
-    Node as ProsemirrorNode,
-} from "prosemirror-model";
+import { Mark, MarkType, NodeType, Schema } from "prosemirror-model";
 import {
     Command,
     EditorState,
@@ -512,32 +506,4 @@ export function splitCodeBlockAtStartOfDoc(
     }
 
     return splitBlock(state, dispatch);
-}
-
-function isSelectionInCodeBlock(
-    state: EditorState
-): { pos: number; node: ProsemirrorNode } | null {
-    const { $from } = state.selection;
-    if ($from.parent.type.name === "code_block") {
-        return { pos: $from.before(), node: $from.parent };
-    }
-    return null;
-}
-
-// Command to open the language dropdown.
-export function openCodeBlockLanguagePicker(
-    state: EditorState,
-    dispatch: (tr: Transaction) => void
-) {
-    const codeBlock = isSelectionInCodeBlock(state);
-    if (!codeBlock) {
-        return false;
-    }
-    const { pos, node } = codeBlock;
-    // Update the node attributes to trigger the language input.
-    const newAttrs = { ...node.attrs, isEditingLanguage: true };
-    if (dispatch) {
-        dispatch(state.tr.setNodeMarkup(pos, undefined, newAttrs));
-    }
-    return true;
 }
